% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{HE}
\alias{HE}
\title{HE function}
\usage{
HE(mHUI = NULL, mHUI_input = NULL, Nx = NULL, Dx = NULL, Mx = Dx/Nx,
  ages = c(0, 1, seq(5, 90, by = 5)), age_interval = c(diff(c(0, 1, seq(5,
  90, by = 5))), 5), axmethod = "midpoint", sex = "male", mxsmooth = TRUE,
  axsmooth = TRUE, radix = 1e+05, verbose = TRUE,
  uncertainty_range = TRUE, alpha = 0.05, actual_death_counts = NULL,
  survival_plots = TRUE, digits = 2)
}
\arguments{
\item{mHUI}{the mean Health Utilities Index, an important argument for this function, can be HRQOL or prevalence of disability.}

\item{mHUI_input}{a character indicator for mHUI. Accept either 'HRQOL' or 'prevalence'.}

\item{Nx}{numeric vector of population exposures by age.}

\item{Dx}{numeric vector of death counts by age.}

\item{Mx}{(optional) numeric vector of central death rates (assumed in the function to be the lifetable m(x)), calculated as deaths/exposure.}

\item{ages}{a numeric vector indicating data age group. (Not an optional argument!), default is 0~1,1~4, 5-year abridged age groups, all the way up to 90+ (A total of 20 age groups).}

\item{age_interval}{defines the number of years in the age interval. Note: the default maximum age group is 90+, for these particular age group, years in interval are also set to 5 by hand, with "age_interval=c(diff(c(0,1,seq(5,90,by=5))),5)".}

\item{axmethod}{either "keyfitz", "schoen", "preston" or "midpoint". Default = "keyfitz", although this is not recommended for abridged ages. See comparisons in axEstimate examples. The user can also supply a numeric vector of a(x) values here (e.g. from a different estimation procedure or from a different population).}

\item{sex}{either "male" or "female" (default). It is only necessary to specify this if "preston" is the axmethod. It does not affect any other lifetable calculations.}

\item{mxsmooth}{logical, default = TRUE. Should the mx vector be smoothed? If TRUE and both Nx and Dx vectors are supplied (the ideal case), smoothing is done using the function Mort1Dsmooth() from Giancarlo Camarda's MortalitySmooth package. In this case, Dx values are smoothed using log(Nx) as an offset, and all other items are the function defaults. If Mx is provided instead of Nx and Dx a loess smoother is used, loess, with span set to .15 for single age data and .4 for 5-year abridged data. If these smoothing procedures are not satisfactory, the user may wish to pre-process the Mx estimate and specify mxsmooth = FALSE, or else leave it rough.}

\item{axsmooth}{logical, default = TRUE. Ignored if mxsmooth = TRUE. Should the a(x) values be calculated from a smoothed M(x) series? In this case, the M(x) series is smoothed within the axEstimate() function for a(x) estimation, but the smoothed M(x) function that was used is not returned. In general, it is better to smooth the M(x) function prior to putting it in this function, because the loess smoother used here has no weights or offset. If this is not possible, loess M(x) smoothing still produces more consistent and less erratic a(x) estimates. If mxsmooth = FALSE and axsmooth = TRUE, the Mx series is only smoothed for use in a(x) estimation, and does not affect any other lifetable calculations that are dependent on Mx.}

\item{radix}{The lifetable starting population at age 0, l(0). default = 1. Other common values are 1000 and 100000, although any value may be given.}

\item{verbose}{logical, default = TRUE. Should informative but possibly annoying messages be returned when the function does something that you might want to know about?}

\item{uncertainty_range}{an argument controlling the output of uncertainty range, default to TRUE. Also, the default range is double-tail!}

\item{alpha}{sets the alpha level for uncertainty range, default to 0.05. Only works when uncertainty_range=TRUE.}

\item{actual_death_counts}{Actual counts of death for each age interval, used in calculating variances of life table functions.}

\item{survival_plots}{(logical) Do you want to get a survival plot, sir?}

\item{digits}{sets the number of decimal places.}
}
\value{
A list of HALE or DFLE and its uncertainty (if uncertainty_range=TRUE).
\describe { HALE or DFLE is a data.frame file in the output list. The df file has the following basic structure, whether uncertainty_range is TRUE or FALSE:
}
   \item{age}{a numeric vector indicating data age group used by the function.}
   \item{age_interval}{the age interval used in the function.}
   \item{ex}{typical lifetable ex. Life remaining life expectancy at age x. e(0) = life expectancy at birth.}
   \item{HealthExpectancy}{The desired Health expectancy indicator, should be either HALE or DFLE, depending on the specification of mHUI}
\describe { when uncertainty_range=TRUE, extra components will be added to the df columns.
 }
}
\description{
This function estimates HALE or DFLE and its uncertainty (if uncertainty_range=TRUE).
HALE will only be estimated when input of mHUI is HRQOL.
DFLE will only be estimated when input of mHUI is prevalence of disability.
mHUI is an important argument for this function, short for mean Health Utilities Index.
The function produces age-specific HALE or DFLE estimates.
But the default age group should be 20 groups, starting from 0~1, 1~4, and then
by 5 years up to 90+. So it can always be altered to suit specific needs in real data analysis.
}
\details{
HE() function relies heavily on LT() function from LifeTable package, developed by Tim Riffe.
Most of the arguments for this function can be referenced to the LifeTable::LT() function.
This dependent package should come with the installation of the current package, if not mannual install via the following steps:
# install.packages("devtools")
library(devtools)
install_github("timriffe/LifeTable", subdir = "LifeTable")
}
\examples{
### Example 1:
Data from Human Mortality Database, as downloaded in May, 2010. www.mortality.org.

########### Example for calculating HALE

##### without uncertainty range

data(UKRmales1965)
HE(mHUI=HRQOL$Male,
                mHUI_input='HRQOL',
                Mx =UKRmales1965$Mx,
                ages =c(0,1,seq(5,90,by=5)),
                age_interval=c(diff(c(0,1,seq(5,90,by=5))),5),
                axmethod = "schoen",
                sex = "male",
                mxsmooth = TRUE,
                axsmooth = TRUE,
                radix = 100000,
                verbose = TRUE,
                digits=2,
                survival_plots=TRUE,
                uncertainty_range=FALSE)


##### with uncertainty range

data(UKRmales1965)
HE(mHUI=HRQOL$Male,
                mHUI_input='HRQOL',
                Mx =UKRmales1965$Mx,
                ages =c(0,1,seq(5,90,by=5)),
                age_interval=c(diff(c(0,1,seq(5,90,by=5))),5),
                axmethod = "schoen",
                sex = "male",
                mxsmooth = TRUE,
                axsmooth = TRUE,
                radix = 100000,
                verbose = TRUE,
                uncertainty_range=TRUE,
                alpha=0.05,
                digits=2,
                survival_plots=TRUE,
                actual_death_counts=UKRmales1965$Dx)

### Example 2: (with uncertainty range)

data(UKR5males1965_disabilityprevalence)
data(UKRmales1965)
HE(mHUI=UKR5males1965_disabilityprevalence$Male,
                mHUI_input='prevalence',
                Mx =UKRmales1965$Mx,
                ages =c(0,1,seq(5,90,by=5)),
                age_interval=c(diff(c(0,1,seq(5,90,by=5))),5),
                axmethod = "schoen",
                sex = "male",
                mxsmooth = TRUE,
                axsmooth = TRUE,
                radix = 100000,
                verbose = TRUE,
                uncertainty_range=TRUE,
                alpha=0.05,
                digits=2,
                survival_plots=TRUE,
                actual_death_counts=UKRmales1965$Dx)

}
\keyword{DFLE}
\keyword{HALE,}
