% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UKR5males1965_disabilityprevalence}
\alias{UKR5males1965_disabilityprevalence}
\title{Example disability data for Ukraine males.}
\format{A data frame with 20 rows and 3 variables:
\describe{
  \item{AgeGroup}{20 age groups, starting from 0~1 and 1~4, then abridged in 5 year interval, up to 90+}
  \item{Male}{Male prevalence rate}
  \item{Female}{Female prevalence rate}

}}
\source{
\url{http://www.mortality.org/} at Human Mortality Database, as downloaded in May, 2010.
}
\usage{
data(UKR5males1965_disabilityprevalence)
}
\description{
5-year adbridged data prevalence rate data.
}
\details{
All three columns are numeric.
}
\references{
Actually, this data is FAKE, only for the purpose of demonstration. It originally comes from the chapter Session 5.1 Calculating DALYs and HALE, page 12.
}
