% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UKRmales1965}
\alias{UKRmales1965}
\title{Example data for Ukraine males, 1965, extracted from the Human Mortality Database.}
\format{A data frame with 20 rows and 4 variables:
\describe{
  \item{age}{20 age groups, starting from 0~1 and 1~4, then abridged in 5 year interval, up to 90+}
  \item{Dx}{Male Death counts}
  \item{Nx}{Male population, so-called exposure population}
  \item{Mx}{central death rates, equal to death/exposure}

}}
\source{
\url{http://www.mortality.org/} at Human Mortality Database, as downloaded in May, 2010.
}
\usage{
data(UKRmales1965)
}
\description{
5-year adbridged data from the HMD for Deaths (count) Population (exposure) and central death rates, death/exposure.
}
\details{
All four columns are numeric. Mx provided is not exactly equal to D(x)/N(x) due to rounding.
}
\references{
Human Mortality Database. University of California, Berkeley (USA), and Max Planck Institute for Demographic Research (Germany). Available at www.mortality.org or www.humanmortality.de (data downloaded on 10 May, 2010).
}
